Error are displayed on uploaded file oversize limit

+++ Form/AdminMediaHandler.php
@@ -27,43 +27,48 @@ class AdminMediaHandler extends AbstractAdminHandler
                 $file = $this->form->get('media')->getData();
                 if (!is_null($file))
                 {
-                    $oldFile = LightCMSUtils::getUploadDir(FALSE) . $media->getFullpath();
-                    if (is_file($oldFile))
-                    {
-                        unlink($oldFile);
-                    }
-                    $filename = LightCMSUtils::getUniqName($file, LightCMSUtils::getUploadDir());
-                    $media->setFullPath(LightCMSUtils::getUploadUrl() . $filename);
-                    $media->setOriginalName($file->getClientOriginalName());
-                    $mimeType = $file->getMimeType();
-                    $media->setMediaType($mimeType);
-                    $file->move(LightCMSUtils::getUploadDir(), $filename);
-                    $mediaType = preg_split('/\//', $mimeType);
-                    if (isset($this->thumbSizes) && $mediaType[0] == 'image')
-                    {
-                        foreach ($this->thumbSizes as $size)
+                    if ($file->getError() == UPLOAD_ERR_OK) {
+                        $oldFile = LightCMSUtils::getUploadDir(FALSE) . $media->getFullpath();
+                        if (is_file($oldFile))
                         {
-                            //@todo: choose crop or resize
-                            LightCMSUtils::cropPicture(
-                                LightCMSUtils::getUploadDir() . $filename,
-                                LightCMSUtils::getUploadDir() . LightCMSUtils::getThumbFilename($filename, $mimeType, $size),
-                                $size['width'],
-                                $size['height']
-                            );
+                            unlink($oldFile);
                         }
+                        $filename = LightCMSUtils::getUniqName($file, LightCMSUtils::getUploadDir());
+                        $media->setFullPath(LightCMSUtils::getUploadUrl() . $filename);
+                        $media->setOriginalName($file->getClientOriginalName());
+                        $mimeType = $file->getMimeType();
+                        $media->setMediaType($mimeType);
+                        $file->move(LightCMSUtils::getUploadDir(), $filename);
+                        $mediaType = preg_split('/\//', $mimeType);
+                        if (isset($this->thumbSizes) && $mediaType[0] == 'image')
+                        {
+                            foreach ($this->thumbSizes as $size)
+                            {
+                                //@todo: choose crop or resize
+                                LightCMSUtils::cropPicture(
+                                    LightCMSUtils::getUploadDir() . $filename,
+                                    LightCMSUtils::getUploadDir() . LightCMSUtils::getThumbFilename($filename, $mimeType, $size),
+                                    $size['width'],
+                                    $size['height']
+                                );
+                            }
+                        }
+//                      $file->getClientSize();
                     }
-//                  $file->getClientSize();
+                    else {
+                        return (FALSE);
+                    }
+                    // New media
+                    if ($media->getId() == 0)
+                    {
+                        $media->setCreatedAt($currentDate);
+                    }
+                    $media->setUpdatedAt($currentDate);
+                    $em = $this->doctrine->getEntityManager();
+                    $em->persist($media);
+                    $em->flush();
+                    return (TRUE);
                 }
-                // New media
-                if ($media->getId() == 0)
-                {
-                    $media->setCreatedAt($currentDate);
-                }
-                $media->setUpdatedAt($currentDate);
-                $em = $this->doctrine->getEntityManager();
-                $em->persist($media);
-                $em->flush();
-                return (TRUE);
             }
         }
         return (FALSE);
